+++ src/Resources/contao/config/config.php
@@ -165,7 +165,7 @@ $GLOBALS['BE_FFL']['tournamentTree'] = 'Schachbulle\ContaoFernschachBundle\Widge
 $GLOBALS['TL_HOOKS']['parseTemplate'][] = array('Schachbulle\ContaoFernschachBundle\Hooks\Newsletter', 'NewsletterTags');
 $GLOBALS['TL_HOOKS']['parseBackendTemplate'][] = array('Schachbulle\ContaoFernschachBundle\Hooks\Template', 'BackendTemplate');
 // Formularversendung prüfen: Beitrittserklärung verarbeiten
-$GLOBALS['TL_HOOKS']['processFormData'][] = array('Schachbulle\ContaoFernschachBundle\EventListener\ProcessFormDataListener', '__invoke');
+$GLOBALS['TL_HOOKS']['processFormData'][] = array('Schachbulle\ContaoFernschachBundle\EventListener\Beitrittsformularpuefung', '__invoke');
 
 /**
  * -------------------------------------------------------------------------